white ball behavior added
the white ball is able to be rotated around the y axis with pressing
left or right arrow buttons. when shift button is pressed with
combination wit left or right arrow then the ball is rotated more
smoothly. when space button is pressed and then released a force
relative to the time of how much the space button was pressed is applied
to the z axis of the ball.

added initial script to the stick game object