Stick behavior completed, some improvements in game physics, some additions in TODO list
Now the cue stick rotate around the white ball and disappears when one
or more balls move.
Some corrections in the game physics and the balls material are made in
order for achieving more realistic billiards game behavior.
Some additions in TODO list for the need of achieving even more
realistic physics are made.