stick animation and audio sounds added
when the the Space key is pressed an animation of the stick is played
like the stick is hitting the white ball and at the end of the animation
a sound is played such of a stick hits a ball. A sound is also played
when two balls collide each other.